requirements.txt
djangorestframework
mongoengine==0.20.0
python-dateutil
Pyjwt
Django==2.2.4
django-cors-headers
nose
Sphinx==1.5.5
sphinx-rtd-theme==0.2.4
openpyxl==2.5.6
django-crontab==0.7.1
Pillow==8.2.0
numpy==1.19.2
Click==7.0
opencv-python==4.4.0.44
six==1.15.0
tensorflow==2.6.0
pandas==1.3.5
tqdm==4.64.1
matplotlib==3.5.3
seaborn==0.12.1
scipy==1.7.3